requirements.txt
astroid==3.3.9
biopython==1.85
black==25.1.0
cfgv==3.4.0
click==8.1.8
dill==0.3.9
distlib==0.3.9
Faker==37.0.2
filelock==3.18.0
flake8==7.1.2
identify==2.6.9
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.4
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.6
pytest==8.3.5
PyYAML==6.0.2
setuptools==75.8.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.2
virtualenv==20.29.3
wheel==0.45.1
